Rework Kernel Cache management
