+++ opencore_legacy_patcher/sys_patch/kernelcache/base/cache.py
@@ -1,0 +1,8 @@
+"""
+cache.py: Base class for kernel cache management
+"""
+
+class BaseKernelCache:
+
+    def rebuild(self) -> None:
+        raise NotImplementedError("To be implemented in subclass")
